Enhance code for better readability
