+++ src/controllers/activity-controller.js
@@ -1,6 +1,5 @@
 const activitiesService = require("../services/activities-service");
-const { uploadFile } = require("../services/s3-service");
-const { deleteTempFile } = require("../services/deleteTempFIles");
+const fileServices = require("../services/fileServices");
 
 
 
@@ -18,23 +17,12 @@ const getAllActivity = async (req, res, next) => {
 
 const createNewActivity = async (req, res, next) => {
   try {
-    //Checking if there's an actual file at the request to decide the value that's gonna be sent as image field 
-    const imageFileCheck = async function () {
-      if (req.file) {
-        const file = req.file;
-        const imageUploadResult = await uploadFile(file)
-        //Deleting image from server once it has been uploaded to aws
-        deleteTempFile(file.path);
-        return imageUploadResult;
-      } else return ("");
-    };
-
-    const image = await imageFileCheck()
+    const image = await fileServices.checkFileAndUpload(req.file)
     const query = req.body;
     console.log(image);
     const activity = await activitiesService.createActivity(
       query,
-      image.Key
+      image
     );
     res.status(200).json(activity);
   } catch (err) {
