+++ src/repositories/activities-repository.js
@@ -13,13 +13,11 @@ const listActivities = async (page) => {
 }
 
 const postActivity = async (activity, imagePath) => {
-    const {name , content} = activity
     console.log(imagePath)
     activity.image = imagePath
     console.log(activity.image)
-
-    const currentActivity = await Activity.create(activity)
     //This will create a new row at the end of Activities table
+    const currentActivity = await Activity.create(activity)   
     return (currentActivity)
 }
